requirements.txt
# AI providers — install at least one
anthropic>=0.25.0
openai>=1.30.0

# Document handling
openpyxl>=3.1.0
python-docx>=1.1.0

# PDF export (requires Microsoft Word installed on Windows/Mac)
docx2pdf>=0.1.8
# PDF text extraction (for Rename tool)
pypdf>=4.0.0
# OCR for scanned PDFs (Rename tool) — also install Tesseract: https://github.com/UB-Mannheim/tesseract/wiki
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
